@@ -1,37 +1,53 @@
 {
-    "name": "cpt-ui",
-    "version": "0.1.0",
-    "private": true,
-    "scripts": {
-        "dev": "next dev",
-        "build": "next build",
-        "start": "next start",
-        "lint": "next lint",
-        "test": "jest",
-        "check-licenses": "license-checker --failOn GPL --failOn LGPL --start ../.."
-    },
-    "dependencies": {
-        "axios": "^1.7.9",
-        "i18next": "^24.2.1",
-        "i18next-browser-languagedetector": "^8.0.0",
-        "i18next-http-backend": "^3.0.2",
-        "next": "15.1.4",
-        "nhsuk-react-components": "^5.0.0",
-        "react": "^18",
-        "react-dom": "^18",
-        "react-i18next": "^15.4.0",
-        "react-intl": "^7.0.1",
-        "sass": "^1.83.4",
-        "uuid": "^11.0.5"
-    },
-    "devDependencies": {
-        "@formatjs/cli": "^6.6.0",
-        "@next/eslint-plugin-next": "^15.1.4",
-        "@types/node": "^22",
-        "@types/react": "^18",
-        "@types/react-dom": "^18",
-        "axios-mock-adapter": "^2.1.0",
-        "babel-plugin-formatjs": "^10.5.35",
-        "eslint-config-next": "15.1.6"
-    }
+  "name": "cpt-ui",
+  "private": true,
+  "version": "0.1.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview",
+    "test": "NODE_NO_WARNINGS=1 jest --no-cache --coverage"
+  },
+  "dependencies": {
+    "aws-amplify": "^6.0.0",
+    "i18next": "^24.2.1",
+    "i18next-browser-languagedetector": "^8.0.0",
+    "i18next-http-backend": "^3.0.1",
+    "nhsuk-frontend": "^8.0.2",
+    "nhsuk-react-components": "^5.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^15.4.0",
+    "react-intl": "^7.0.1",
+    "react-router-dom": "^6.22.3"
+  },
+  "devDependencies": {
+    "@swc/core": "^1.10.11",
+    "@swc/jest": "^0.2.37",
+    "@testing-library/jest-dom": "^6.6.2",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^20.11.17",
+    "@types/react": "^18.2.64",
+    "@types/react-dom": "^18.2.21",
+    "@typescript-eslint/eslint-plugin": "^8.22.0",
+    "@typescript-eslint/parser": "^8.22.0",
+    "@vitejs/plugin-react": "^4.2.1",
+    "eslint": "^9.16.0",
+    "eslint-plugin-jest": "^28.11.0",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest-watch-typeahead": "^2.2.2",
+    "sass": "^1.70.0",
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.2.2",
+    "vite": "^5.1.6"
+  }
 }
